Rename modules. vc -> credential
