+++ bindings/wasm/src/vc.rs
@@ -1,6 +1,6 @@
 use identity_core::{
     common::{OneOrMany, Url},
-    vc::{Credential, CredentialBuilder, CredentialSubject, VerifiableCredential as VC},
+    credential::{Credential, CredentialBuilder, CredentialSubject, VerifiableCredential as VC},
 };
 use wasm_bindgen::prelude::*;
