+++ examples/credential/src/main.rs
@@ -4,17 +4,17 @@
 use identity_core::{
     common::{Url, Value},
     convert::{FromJson as _, ToJson as _},
+    credential::{Credential, CredentialBuilder, CredentialSubject, VerifiableCredential},
     did_doc::MethodScope,
     did_url::DID,
     json,
-    vc::{Credential, CredentialBuilder, CredentialSubject, VerifiableCredential},
 };
 use identity_iota::{
     client::Client,
+    credential::{CredentialValidation, CredentialValidator},
     crypto::KeyPair,
     did::IotaDocument,
     error::Result,
-    vc::{CredentialValidation, CredentialValidator},
 };
 
 // A helper function to generate and new DID Document/KeyPair, sign the
